requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
opencv-python==4.8.1.78
Pillow==9.5.0
numpy==1.21.6
pymongo==4.6.1
insightface==0.7.3
tqdm==4.66.1
